Finally, into th main file, I used | to break the lines into 4 parts, Created an elpoyee object, and displayed the employeed ID, name and gross pay

+++ src/main/java/com/pluralsight/Employee.java
@@ -12,7 +12,7 @@ public class Employee {
 
 
     //Constructors
-    public Employee(int employeeId, String name, int hoursWorked, double payRate) {
+    public Employee(int employeeId, String name, double hoursWorked, double payRate) {
         this.employeeId = employeeId;
         this.name = name;
         this.hoursWorked = hoursWorked;
